third_party/WebKit/LayoutTests/fast/forms/radio/radio-input-keyboard-navigation.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<style>
	#test1 {
		max-height: 65px;
		overflow: auto;
		width: 100px;
		height: 65px;
	}
	input[type="radio"] {
		display: block;
	}
</style>
</head>
<body>
<div id="test1">
	<input type="radio" name="foo" checked>
	<input type="radio" name="foo">
	<input type="radio" name="foo">
	<input type="radio" name="foo">
	<input type="radio" name="foo">
	<input type="radio" name="foo">
</div>
<script>
	test(function() {
		document.querySelector("input").focus();
		assert_own_property(window, "eventSender");
		for (var i = 0; i < 5; i++) {
			eventSender.keyDown("ArrowDown");
		}
		assert_greater_than(document.getElementById("test1").scrollTop, 0);
	}, "Parent container should be scrolled while navigating through radio inputs");
</script>
</body>
</html>
